[JAX Runner] Refactor jax runner to delegate KV Cache handling logic. [3/N]
